requirements.txt
PySide6>=6.6
Pillow>=10.0
mss>=9.0
pywin32>=306; sys_platform == "win32"
